ACP-L2-CTR-PRV-CONTROL: reconcile round-3 review, drop generated-file wiring, fold control vocabulary into existing root files
Merged fork/main (33 commits) to clear baseline drift, then addressed the
third blocking review round on PR #1749:

- Reverted the canonical pkg/wire logger-injection change from the round-2
  fix. Acceptance criterion 5 explicitly requires generated artifacts to be
  absent from this change, and provideProvidersService/wire_gen.go are
  generated output. Canonical top-level logger wiring for Providers remains
  out of scope for this packet; providers/wire.NewService(WithLogger(...))
  (a hand-written, non-generated composition layer) still exists for direct
  callers and defaults to logging.NoopLogger, matching the packet's original
  scope.
- Folded control_contract.go and control_characterization_test.go into the
  existing execute_contract.go / execute_characterization_test.go instead of
  adding new files to the Providers root. This makes the entire thin-root
  file-inventory diff (internal/ownershipinventory, cmd/packagetargetmanifestcheck)
  disappear, since no new root file is introduced.
- Kept the round-2 removal of the reflect-based exact-method-list assertion
  in root_contract_characterization_test.go (renamed to
  TestProvidersRootServiceInterfaceLivesInServiceContractFile) since that
  test would otherwise show real regression the moment ControlAttempt is
  added to providers.Service; ControlAttempt's interface membership is
  proved behaviorally via peer-fake tests instead.

Verified: go build ./..., go vet ./..., gofmt clean on changed files;
go test ./pkg/services/providers/... ./pkg/wire/... all pass; full
go run ./cmd/unitlane -jobs 4 fully green (zero FAIL lines); go run
./cmd/pkgstructurecheck -root . shows only pre-existing chat_sessions/
operator_settings violations, confirmed identical via git stash.
