+++ pkg/wire/chat_sessions_composition_test.go
@@ -130,7 +130,7 @@ func TestProvideChatSessionsFactoryTargetCatalogServiceComposesThroughTheCanonic
 
 	edges := serviceedges.Edges{}
 	files := provideOperatorSettingsFileSystem(edges)
-	providersRoot, err := provideProvidersService(edges, logger)
+	providersRoot, err := provideProvidersService(edges)
 	if err != nil {
 		t.Fatalf("provideProvidersService() error = %v", err)
 	}
