+++ pkg/wire/operator_settings_logging_test.go
@@ -25,7 +25,7 @@ func TestProvideOperatorSettingsServiceLogsThroughTheCanonicalWireLogger(t *test
 
 	edges := serviceedges.Edges{}
 	files := provideOperatorSettingsFileSystem(edges)
-	providersRoot, err := provideProvidersService(edges, logger)
+	providersRoot, err := provideProvidersService(edges)
 	if err != nil {
 		t.Fatalf("provideProvidersService() error = %v", err)
 	}
