+++ pkg/wire/runtime_provider_bindings.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
 	operatorsettings "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 	"github.com/portpowered/infinite-you/pkg/services/providers"
@@ -18,9 +17,8 @@ func provideRuntimeProviderBindings(
 	edges serviceedges.Edges,
 	integrations []operatorsettings.ACPIntegration,
 	runner workers.CommandRunner,
-	logger logging.Logger,
 ) (workers.ProviderRegistry, providers.Service, workerswire.ProviderRegistryRebinder, error) {
-	runtimeProviders, err := provideConfiguredProvidersService(edges, logger, integrations, runner)
+	runtimeProviders, err := provideConfiguredProvidersService(edges, integrations, runner)
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("construct runtime Providers service: %w", err)
 	}
@@ -29,7 +27,7 @@ func provideRuntimeProviderBindings(
 		return nil, nil, nil, fmt.Errorf("construct runtime provider registry: %w", err)
 	}
 	rebinder := workerswire.ProviderRegistryRebinder(func(reboundRunner workers.CommandRunner) (workers.ProviderRegistry, providers.Service, error) {
-		reboundProviders, rebuildErr := provideConfiguredProvidersService(edges, logger, integrations, reboundRunner)
+		reboundProviders, rebuildErr := provideConfiguredProvidersService(edges, integrations, reboundRunner)
 		if rebuildErr != nil {
 			return nil, nil, rebuildErr
 		}
