+++ pkg/wire/system_initialization_composition_test.go
@@ -106,7 +106,7 @@ func TestProvideSystemInitializationServiceComposedInitializeCreatesThenSkipsPac
 
 	edges := serviceedges.Edges{}
 	files := provideOperatorSettingsFileSystem(edges)
-	providersRoot, err := provideProvidersService(edges, logging.NoopLogger{})
+	providersRoot, err := provideProvidersService(edges)
 	if err != nil {
 		t.Fatalf("provideProvidersService() error = %v", err)
 	}
